pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
# Check https://flit.readthedocs.io/en/latest/pyproject_toml.html for all available sections
name = "ansys-mapdl-core-tutorials"
version = "0.1.dev0"
description = "Tutorials for PyMAPDL package."
readme = "README.rst"
requires-python = ">=3.7"
license = {file = "LICENSE"}
authors = [
    {name = "Ansys, Inc.", email = "pyansys.support@ansys.com"},
]
maintainers = [
    {name = "PyAnsys developers", email = "pyansys.support@ansys.com"},
]
dependencies = [
    "ansys-mapdl-core>=0.63.2",
    "jupyterlab>=3.2.8",
    "matplotlib==3.6.0",
    "pandas==1.5.0",
    "plotly==5.10.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[tool.flit.module]
name = "ansys.mapdl.core.tutorials"

[project.urls]
Source = "https://github.com/pyansys/spacecon-spain-2022"
